+++ _suitkaise_wip/fdl/_int/_private_api.py
@@ -1,4 +1,0 @@
-# API for internal suitkaise use ONLY!!!
-
-# this doesn't wrap around core logic, but uses it directly
-# use this file to gather all core logic for use inside other suitkaise modules