+++ _suitkaise_wip/fileconv/_int/base/formats.py
@@ -1,9 +1,0 @@
-
-
-class fmt(Enum):
-
-    BOLD = "bold"
-    ITALIC = "italic"
-    CODE = "code"
-    UNDERLINE = "underline"
-    STRIKETHROUGH = "strikethrough"