+++ _suitkaise_wip/fileconv/_int/base/parser.py
@@ -1,3 +1,0 @@
-class Parser(ABC):
-
-    def __parse__(self, text: str) -> TextFile: