+++ _suitkaise_wip/fileconv/_int/base/text_file.py
@@ -1,3 +1,0 @@
-class TextFile:
-
-    # main object that parsers create, and converters receive