+++ _suitkaise_wip/fileconv/_int/pieces/ascii.py
@@ -1,5 +1,0 @@
-from ..base.piece import Piece
-
-
-class AsciiPiece(Piece):
-    pass